overhauled game entites to use floats for their physical properties

+++ Engine/Dude.h
@@ -8,14 +8,14 @@ public:
 	void Update(const Keyboard& kbd);
 	void ClampToScreen();
 	void Draw(Graphics& gfx) const;
-	int GetX() const;
-	int GetY() const;
-	int GetWidth() const;
-	int GetHeight() const;
+	float GetX() const;
+	float GetY() const;
+	float GetWidth() const;
+	float GetHeight() const;
 private:
-	int x = 400;
-	int y = 300;
-	int speed = 2;
-	static constexpr int width = 20;
-	static constexpr int height = 20;
+	float x = 400.0f;
+	float y = 300.0f;
+	float speed = 2.0f;
+	static constexpr float width = 20.0f;
+	static constexpr float height = 20.0f;
 };
